First pass on conversion to ts
