+++ package.json
@@ -5,32 +5,43 @@
     "node": ">=10.0.0"
   },
   "description": "",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node /src/index.js",
-    "start:dev": "nodemon --watch src src/index.js",
+    "start:dev": "nodemon --watch src dist/index.js",
     "prettier": "prettier --write \"src/**/*.js\"",
-    "lint": "eslint src"
+    "lint": "eslint src",
+    "build": "tsc",
+    "prestart": "npm run build",
+    "start": "node ."
   },
   "author": "Andrew Nguyen",
   "license": "ISC",
   "dependencies": {
+    "@types/bcrypt": "^3.0.0",
+    "@types/express": "^4.16.1",
+    "@types/graphql": "^14.2.0",
+    "@types/jsonwebtoken": "^8.3.2",
+    "@types/lodash": "^4.14.123",
+    "@types/mongoose": "^5.3.26",
     "apollo-server-express": "^2.4.8",
     "bcrypt": "^3.0.6",
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "express-graphql": "^0.7.1",
     "graphql": "^14.2.1",
+    "i": "^0.3.6",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.11",
-    "mongoose": "^5.5.1"
+    "mongoose": "^5.5.1",
+    "npm": "^6.9.0"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-plugin-node": "^8.0.1",
     "nodemon": "^1.18.11",
-    "prettier": "^1.17.0"
+    "prettier": "^1.17.0",
+    "typescript": "^3.4.3"
   }
 }
