+++ src/app/users/model.ts
@@ -1,6 +1,6 @@
-const mongoose = require('mongoose')
+import mongoose = require('mongoose')
 
-const UserSchema = mongoose.Schema(
+const UserSchema = new mongoose.Schema(
 	{
 		name: String,
 		email: String,
