+++ src/utils/constants.d.ts
@@ -1,0 +1,1 @@
+export const SALT_ROUNDS = 12
