--- src/utils/constants.js
@@ -1,2 +1,0 @@
-module.exports.SALT_ROUNDS = 12
-module.exports.JWT_SECRET = process.env.JWT_SECRET
